requirements.txt
alabaster==1.0.0
altgraph==0.17.4
astroid==3.3.9
babel==2.17.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
dill==0.3.9
docutils==0.21.2
flake8==7.1.2
google-api-core==2.23.0
google-api-python-client==2.153.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
httplib2==0.22.0
idna==3.10
imagesize==1.4.1
isort==6.0.1
Jinja2==3.1.6
loguru==0.7.3
MarkupSafe==3.0.2
mccabe==0.7.0
mypy==1.15.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.2
pdoc==15.0.1
pefile==2023.2.7
platformdirs==4.3.6
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pyflakes==3.2.0
Pygments==2.19.1
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
pylint==3.3.5
pyparsing==3.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
pywin32-ctypes==0.2.3
requests==2.32.3
requests-oauthlib==2.0.0
roman-numerals-py==3.1.0
rsa==4.9
setuptools==75.5.0
snowballstemmer==2.2.0
Sphinx==8.2.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
tomlkit==0.13.2
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
